3.3: Add interfaces & remove premature returning
- Premature returning was buggy and didn't stand up to real-world
testing. I couldn't figure out how to debug some weird issues related to
it, which really seemed like it was somehow corrupting Zend VM memory,
so I opted to remove the whole feature. THIS BREAKS BACKWARDS
COMPATIBILITY, but nobody was using the couple releases it was an
available feature in so I don't care.

- Promises now support rejection with all Throwable classes, including
PHP errors, and not Exceptions. This extends compatibility, but you must
beware of this change in your consumers, since if you're really meaning
to catch everything, now PHP errors can make it through.

- Major classes now have interfaces, enabling the ability for users to
create their own implementations.

- Some unfinished code from 2022 was finally removed from the codebase.